fix: remove redundant install.ts and reorganize
