+++ create-quasar/templates/app-extension/ae-ts/install-script/app-extension/src/templates/README.md
